Implement the new ManifestRawStorage and a MappedRawStorage interface
